pyproject.toml
[project]
name = "fbref-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "httpx-curl-cffi>=0.1.3",
    "jsonlines>=4.0.0",
    "lxml>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
]
